Update README.md
Program Description:

"This Python program allows the user to ask a question, and the system responds randomly with one of two options: 'truth' or 'bindolery'. It's a simple game that simulates an oracle providing fun and unexpected answers. The program is distributed under the open-source GPL 3.0 license and can be freely modified and redistributed."